added non-generic IFakeOptions
Only supports ConfigureFake right now.
No externally-visible change in behaviour.
