+++ Source/FakeItEasy.Net35/FakeItEasy.Net35.csproj
@@ -425,6 +425,9 @@
     <Compile Include="..\FakeItEasy\Creation\FakeObjectCreator.cs">
       <Link>Creation\FakeObjectCreator.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\Creation\IFakeOptions.of.T.cs">
+      <Link>Creation\IFakeOptions.of.T.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\Creation\IProxyOptions.cs">
       <Link>Creation\IProxyOptions.cs</Link>
     </Compile>
