+++ Source/FakeItEasy/Creation/IFakeOptions.cs
@@ -1,75 +1,20 @@
 namespace FakeItEasy.Creation
 {
     using System;
-    using System.Collections.Generic;
-    using System.Diagnostics.CodeAnalysis;
-    using System.Linq.Expressions;
-    using System.Reflection.Emit;
     using FakeItEasy.Configuration;
 
     /// <summary>
     /// Provides options for generating fake object.
+    /// Has reduced functionality when compared to <see cref="IFakeOptions{T}"/>,
+    /// which should be used when the type of the fake being created is known.
     /// </summary>
-    /// <typeparam name="T">The type of fake object generated.</typeparam>
-    public interface IFakeOptions<T>
+    public interface IFakeOptions
         : IHideObjectMembers
     {
         /// <summary>
-        /// Specifies arguments for the constructor of the faked class.
-        /// </summary>
-        /// <param name="argumentsForConstructor">The arguments to pass to the constructor of the faked class.</param>
-        /// <returns>Options object.</returns>
-        IFakeOptions<T> WithArgumentsForConstructor(IEnumerable<object> argumentsForConstructor);
-
-        /// <summary>
-        /// Specifies arguments for the constructor of the faked class by giving an expression with the call to
-        /// the desired constructor using the arguments to be passed to the constructor.
-        /// </summary>
-        /// <param name="constructorCall">The constructor call to use when creating a class proxy.</param>
-        /// <returns>Options object.</returns>
-        [SuppressMessage("Microsoft.Design", "CA1006:DoNotNestGenericTypesInMemberSignatures", Justification = "This is by design when using the Expression-, Action- and Func-types.")]
-        IFakeOptions<T> WithArgumentsForConstructor(Expression<Func<T>> constructorCall);
-
-        /// <summary>
-        /// Specifies that the fake should delegate calls to the specified instance.
-        /// </summary>
-        /// <param name="wrappedInstance">The object to delegate calls to.</param>
-        /// <returns>Options object.</returns>
-        IFakeOptionsForWrappers<T> Wrapping(T wrappedInstance);
-
-        /// <summary>
-        /// Specifies that the fake should be created with these additional attributes.
-        /// </summary>
-        /// <param name="customAttributeBuilders">The attributes to build into the proxy.</param>
-        /// <returns>Options object.</returns>
-        IFakeOptions<T> WithAdditionalAttributes(IEnumerable<CustomAttributeBuilder> customAttributeBuilders);
-
-        /// <summary>
-        /// Sets up the fake to implement the specified interface in addition to the
-        /// originally faked class.
-        /// </summary>
-        /// <param name="interfaceType">The type of interface to implement.</param>
-        /// <returns>Options object.</returns>
-        /// <exception cref="ArgumentException">The specified type is not an interface.</exception>
-        /// <exception cref="ArgumentNullException">The specified type is null.</exception>
-        [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Implements", Justification = "Would be a breaking change, might be changed in a later major version.")]
-        IFakeOptions<T> Implements(Type interfaceType);
-
-        /// <summary>
-        /// Sets up the fake to implement the specified interface in addition to the
-        /// originally faked class.
-        /// </summary>
-        /// <typeparam name="TInterface">The type of interface to implement.</typeparam>
-        /// <returns>Options object.</returns>
-        /// <exception cref="ArgumentException">The specified type is not an interface.</exception>
-        [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Implements", Justification = "Would be a breaking change, might be changed in a later major version.")]
-        [SuppressMessage("Microsoft.Design", "CA1004:GenericMethodsShouldProvideTypeParameter", Justification = "Used to provide a strongly typed fluent API.")]
-        IFakeOptions<T> Implements<TInterface>();
-
-        /// <summary>
         /// Specifies an action that should be run over the fake object for the initial configuration (during the creation of the fake proxy).
         /// </summary>
-        /// <param name="action">An action to perform.</param>
+        /// <param name="action">An action to perform on the Fake.</param>
         /// <returns>Options object.</returns>
         /// <remarks>
         /// <para>
@@ -80,6 +25,6 @@ namespace FakeItEasy.Creation
         /// the <c>action</c> will be called more than once, so it should be side effect-free.
         /// </para>
         /// </remarks>
-        IFakeOptions<T> ConfigureFake(Action<T> action);
+        IFakeOptions ConfigureFake(Action<object> action);
     }
 }
